New version of Baglight with Dimmer control
added second switch wired like the first but to pin 10 This allows the lights brightness to swtich between a regular brightness (85) and max brightness. I also added a lantern mode for the light this is just a constant on white. previous to this all the modes were some sort of blinking or animations,. 
cylon eyes are now only 4 pixels wide and are constant red not red and blue. 

I also re-ordered the modes as follows
1 - white blinking
2 - white constant on
3 - Red and White blink
4 - Red and White left right blink
5 - Red and white crawl animation
6 - Cylon Eye - red eye on white background
7 - rainbow animation